Code cleaned and commented
