+++ app/src/main/java/com/github/freeman/bootcamp/SettingsProfileActivity.kt
@@ -4,7 +4,6 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.graphics.Bitmap
-import android.graphics.BitmapFactory
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -40,14 +39,12 @@ import coil.compose.rememberAsyncImagePainter
 import com.github.freeman.bootcamp.auth.FirebaseAuthActivity
 import com.github.freeman.bootcamp.utilities.firebase.FirebaseUtilities
 import com.github.freeman.bootcamp.ui.theme.BootcampComposeTheme
-import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.ktx.storage
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import java.io.ByteArrayOutputStream
 
 /**
  * Activity that shows up the settings and your profile
@@ -367,8 +364,7 @@ private fun prepareOptionsData(context: Context) {
             title = "Manage Account",
             subTitle = "Sign in or sign out from your Google account",
             clickAction = {
-                context.startActivity(Intent(context, FirebaseAuthActivity::class.java)
-                )
+                context.startActivity(Intent(context, FirebaseAuthActivity::class.java))
             }
         )
     )
